+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 IMPROVEMENTS
 
+* WIRBELSTURM-5: `$MYDIR` in AWS scripts should be `$MY_DIR` (thanks pbwilmot)
 * WIRBELSTURM-2: `bootstrap` now supports the `--skip-ruby` parameter.  If set, `bootstrap` will skip the installation
   of Ruby/rvm/bundler/gems. (thanks bzz)
 * Environment `default_environment`: enforce setting timezone to 'UTC'